True fix for salt-and-pepper noise
It turns out that a much larger distance is needed to offset the
limited precision of 32-bit floating point. 0.0001 seems kind of big.
This means that the scale of scenes that we can display is limited to
a certain range, although I don't think will be an issue. I wonder how
much smaller a distance we could use if we used 64-bit numbers instead
of 32-bit? I've been using 32 because I assumed the calculations would
be faster, but I haven't actually tested this.