change shape.material() to return a reference
I was trying to set values directly on the returned value, which was
a clone of the one inside of the shape itself. This way we make it
obvious that we have to clone and re-set the value. Although there are
some outstanding TODO's that will make this look a lot nicer.