Non-compiling attempt
Groups have shapes children, which in turn have groups as parents...
Who owns who? I backed myself into a corner on this one. I can't just
pass `self` to an `Rc` constructor, because then the Rc would own the
current object, which I'm sure is bad and the compiler is going to stop
me.