Research on doing parent/child linking
Once again, I thought of borrowing a technique from @undo76:

https://github.com/undo76/raytracer-rust/blob/master/core/src/shape.rs

Specifically the way that a shape's parent is stored. This is what I
understand of the method:

AtomicPtr contains a mutable pointer to some piece of data that is
itself not owned. Mutable pointers are unsafe in general, and are thus
stored in UnsafeCell's internally.

The default value for a pointer is null, and AtomicPtr.load will return
None for this case. AtomicPtr::new is totally safe.

The unsafety shows in the load and store methods, which are marked
unsafe. storing and loading a shared mutable pointer is unsafe because
no gaurantees are given regarding what other clients are doing with the
contents of these pointers. The contents could be overwritten and lost,
invalidating the pointer.

I think it's safe for our use case because 1) we will be setting the
parent object once and then never changing it and 2) the GroupShape
objects will own the children shapes that contain the pointers to them,
meaning that the children will be dropped before the parents and that
therefore there is no chance that the parent pointers will be
invalidated by the GroupShape object being dropped.

That said, I would rather have these invariants enforced somehow, since
we're わざわざ learning Rust, a language that is supposed to be
completely memory safe, for this project.