abandoning attempt to change to safe references
I thought maybe we could use weak refs to refer to parents and Rc for
children. The problem is that all children of a group need to be given
the same weak ref, meaning that the parent GroupShape needs to keep
its own weak reference around. That can't be directly in the
constructor; the client would be required to create an Rc containing the
new GroupShape and then pass a downgraded ref to the GroupShape object,
which could then store it and pass it to its children. This would mean
having a `self_ref: Option<Weak<Self>>` field and then in the add_child
method we would also have to check that `self_ref` had been set or else
give the client an error of some kind. So it could work and it would be
safe, but it would be a tad heavy.

Actually, since it's not even compiling right now, I think it would
actually require an Option<Weak<RefCell<Self>>> because the client needs
to be able to borrow mutably. But then the borrow checker is running at
runtime, okay this just sucks.