(Failing) attempt to implement world_to_object
Because of ownership rules, it doesn't seem possible to test this
simply. I can't get a grandchild node without unsafe casting it to a
GroupShape (which I don't even know how to do in Rust). I guess there's
no dynamic dispatch of any kind, because GroupShape's methods don't get
called if a variable is declared as just a dyn Shape.

I tried returning references to the added children in `add_child`,
so the test looked like this:

    let s = g2.add_child(Box::new(s));
    g1.add_child(Box::new(g2));

    let p = s.as_ref().world_to_object(point!(-2, 0, -10));
    assert_eq!(p, point!(0, 0, -1));

but no dice! Returning a reference from a method constitutes a borrow
of the object it was called on, which I guess is a good call by Rust.

So I guess really the only way to go about this correctly is to have
a single object to own the whole tree, which means using an arena
of some kind.