Propagate GroupShape transforms to children
Further documentation and cleanups are still todo, but the basic idea
is that we can avoid annoying borrowing issues in Rust and also probably
make everything more efficient by propagating group transforms directly
to the children and not looping recursively through parents to convert
from world to object co-ordinates. This is fewer matrix multiplications
and also removes the need to have a parent pointer. The book-keeping for
group takes a little more logic, but it is definitely easier than
anything else given the borrow checker's requirements.