Fix bug in group transform propagation
We were multiplying the group's transform inverse on the wrong side.
Finding this drove me a little crazy because I thought there was some
linear algebra thing that was allowing the transform propagation to work
but that prevented the normal from working correctly because of the math
not working out. You can see all of the println statements I was using
to track this down.

The test from the book was not sufficient to catch this case, so that's
been updated. Whew! Now we can move on with the normal.