Cache matrix inverses in camera
Installd cargo-flamegraph (it's fantastic!) and found that a lot of time
was being spent in the matrix inverses in `ray_for_pixel`. Once these
were cached, reflect_refract went from 120s to 50s! When I set the
recursion depth to just 5, it only takes 5 seconds, and it looks exactly
the same!