Rectangle light jitter sampling working
Barely... this is very bad design and needs to be redone. I started by
requiring the client to pass an Rng implementation to RectangleLight,
but moved away from that because there wasn't a clear way to specify
the exact values to be returned during testing. There's a mock Rng
in rand_core that returns some step function, but you can't control the
f32 values that these get turned into. The plain iterator turned out to
be really gross to work with because calling `next()`

Looking at the docs some more, I think the correct thing to use is a
`Distribution<f32>`; we can make the `sample` method return test data,
and then we just always use `thread_rng` to sample from it.

I've also added the `derivative` crate to make working with auto-derive
a little bit easier.