Replace intensity_at test
with the new test from the website that uses a hardcoded jitter
sequence. Also use Iterator.cycle to ensure that any future tests won't
panic on an unwrap(); this makes it easier to write tests because we
don't have to count exactly the number of samples we need and pad the
input vector.