Ensure that default base shapes have real object IDs
Previously it was possible to create a BaseShape with the unique ID set to 0;
this was because ID's were simply `usize`, which has a default value of 0.

Move the object ID logic into its own file. Expose a struct instead of just a
`usize`, and implement sane Default and Debug traits for it. Make the id private
so that no client can mutate it. Use this new struct in BaseShape.