Add normalization for OBJ file vertices
Once the first face is encountered, disallow reading any further
vertices and normalize all of the ones that have been read in so far. By
"normalize" I mean to scale down to -1/1 bounds, centered at the origin
and the same size as the regular cubes we are used to dealing with.

Update other tests to not fail; new tests still need to be added to
check that normalization is done correctly.

The book did mention that this was a good idea, but it was discussed on
the forums and I modeled this implementation after the pseudocode
provided by ascotti:
https://forum.raytracerchallenge.com/thread/27/triangle-mesh-normalization