Much improved aabb intersection via tavianator
This code should be more performant (though it only seemed to save about
10% on the one test I did) because it avoids repeated divides and
completely removes normal branching from the intersection code.

There was also a bug in the book's intersection code: we have to check
that the ray intersection is in the correct direction! Otherwise the
tracer would end up showing a mirror image of objects that are actually
behind the camera.

Source:
https://tavianator.com/fast-branchless-raybounding-box-intersections/