remove get_children from Shape trait
It didn't belong there. The only reason it was there was for allowing
tests to access the method on generic Shape trait objects that we knew
were actually GroupShapes. Placing it in the trait also cluttered the
API, since we had to return an Optional instead because all other shape
types had no children.

The alternative is to use downcast_ref to get a typed GroupShape object,
which we know has children. This looks complex when used, but is
reserved solely for the tests.