Don't wrap single groups in groups in obj parser
There are two test cases because originally I had assumed that the
default group ("") always existed, but this is not the case. Return the
one and only existing group from `take_all_as_group`, whether it is
named or not. Add a test cases for both paths.

Also note that the test OBJ I used is the one that I narrowed down from
the low-res teapot when trying to figure out the solution to #5. I have
bonded with it, so I thought it deserved a place :)