Working dragon demo from website
Regarding the map vs. for-loop approach for generating the dragon
displays:

1) The error had nothing to do with lifetimes. It turns out that
`Box<dyn Trait>` is always short for `Box<dyn Trait + 'static>`. See
https://users.rust-lang.org/t/box-with-a-trait-object-requires-static-lifetime/35261/2:

> ... the trait object / type erased element has a : 'static bound
(meaning that it is not borrowing any local, and thus the value is
guarantee not to dangle at any time, which makes it usable for as long
as we wish.

2) The problem was that the `map` step produced `GroupShape` objects,
and the type checker wanted `dyn Shape` objects. But you can't map
straight to `dyn Shape` since it's not size, either. You can give the
compiler a hint at the map to box step by saying `Box::new(x) as _`;
the type checker then figures out that you want a `Box<dyn Shape>`
instead of a `Box<GroupShape>`. Honestly I'm not sure why the compiler
couldn't figure that out without the `as _`. An alternative would be
to annotate the closure passed to `map` with the expected return type.

See https://users.rust-lang.org/t/cannot-extend-a-vec-of-trait-objects/28129: