Planar texture mapping
Note that the `%` operator in Rust is actually a remainder, and for negative
numbers will also return a negative number. We use the `rem_euclid` method
instead to pass the book's tests. It may be worth revisiting this later; it's
slightly less performant because of the `if` branch in the method. For now,
though, we don't know if there will be a requirement somewhere that u and v
cannot be negative.

It's actually kind of interesting to note that different programming languages
have different definitions of the % operator:
https://en.wikipedia.org/wiki/Modulo_operation#In_programming_languages. Since
the book doesn't have any tests where the dividend is negative, it is a little
bit under-specified (again, I'm not sure that it matters). But we do know that
the author used either a language which always returns a positive number (
unlikely, these are rare) or one that returns the sign of the dividend (more
likely, especially since we know the author also likes C).