Cylinder texture mapping
There is a typo in the bonus chapter: the azimuth should be calculated
using x and z, not x and y, which gives interesting but wrong results.

Also, it's not explained, but the author's example cylinder has its y
scaled by PI to get proper squares on the cylinder because the v
coordinate varies betwen 0 and 1 and the u coordinate varies between 0
and pi. I wish there were an easy way to fix this.