Find which cube face a point belongs to
I hesitated with this implementation for a second because I wasn't sure
if floating point numbers would always have x and -x pairs, since the
granularity is not consistent across the real numbers. However, this is
totally fine because floating point numbers store the sign in its own
bit. We could have possibly run into an edge case if we were using
integers instead, since the positive and negative range are not the
same.