Explain normal_to_world logic
Whew! That's basically a blog post in the code, minus pictures. I used a peach
for the example instead of the obvious apple because in Germany now they have
these donut peaches, which make for a great non-uniform scaling example. I guess
it won't make too big a difference for some readers; Gerard told me that in the
Philippines, neither apples nor peaches were a thing :)

I took most of my explanation from
https://paroj.github.io/gltut/Illumination/Tut09%20Normal%20Transformation.html.
There are also a answers in SE showing that this formula is algebraically
provable, but I couldn't quite follow the proofs:
https://stackoverflow.com/questions/13654401/why-transforming-normals-with-the-transpose-of-the-inverse-of-the-modelview-matr
https://computergraphics.stackexchange.com/questions/1502/why-is-the-transposed-inverse-of-the-model-view-matrix-used-to-transform-the-nor