Abandoning for now
The idea was that we could speed up the here_be_dragons rendering by
not requiring the copying of a material to all 21k triangles of each
mode. I thought we could simply store a single material override in
GroupShape and pass it on in an IntersectionList object as an override.
However, there doesn't seem to be a correct behavior for when multiple
subgroups in a group are intersected by a ray. This also just makes the
API messier everywhere. I think in the long run the better thing to do
will be to create a custom TriangleMesh class that stores triangle data
and keeps only one copy of the material for all of the triangles.