Don't sort the intersections in GroupShape
The intersections are all sorted in World, so the computation is redundant.
They do have to be sorted for the tests to pass as-is, though, so place the
sorting in the test method itself. Another way to do this would be to add the
intersections to a hash set and then compare the set of expected and actual
intersections for equality. That's how I would do it in Java.