Convert into cargo workspace
Move all of the demos into their own demo project. Now we can open
just the lib folder or just the demos folder with VS Code, and every
time it recompiles it will not recompile both folders. It saves a little
bit of time; hopefully in the long run it won't be too annoying to have
only one subfolder open.